Backward compatibility fixes for 32-bit PowerPC (GCC 4.0, Mac OS X 10.3/10.4)
- *stdout = *shm_fp instead of stdout = shm_fp (not an lvalue on old macOS)
- MAP_ANONYMOUS compat define for older BSD/macOS MAP_ANON
- (unsigned long long)1 << 40 instead of (size_t)1 << 40 (UB on 32-bit)
- Initialize is_fof to suppress uninitialized variable warning

+++ ladr/memory.c
@@ -519,7 +519,7 @@ void *safe_malloc(size_t n)
     return NULL;
 
   /* Catch negative-int-cast-to-size_t and other overflow bugs. */
-  if (n > (size_t)1 << 40) {
+  if (n > (unsigned long long)1 << 40) {
     fprintf(stderr,
             "safe_malloc: absurd allocation of %lu bytes "
             "(probable integer overflow)\n", (unsigned long)n);
@@ -705,7 +705,7 @@ void *safe_malloc(size_t n)
     return NULL;
 
   /* Catch negative-int-cast-to-size_t and other overflow bugs. */
-  if (n > (size_t)1 << 40) {
+  if (n > (unsigned long long)1 << 40) {
     fprintf(stderr,
             "safe_malloc: absurd allocation of %lu bytes "
             "(probable integer overflow)\n", (unsigned long)n);
